cleanup and add v sense to 3rd adc inj channel
